rtlwifi_new: Add guard for undefined pointer for get_available_desc()
This callback routine pointer was added to driver rtl8192ee to help
fix the stability of the connection. One problem with that fix is that
this pointer is undefined for the rest of the drivers. To prevent
crashes, we need to test for a non-NULL pointer before we call the
routine.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>