rtlwifi_new: Do not call is_tx_desc_closed() if packet was a beacon
When a beacon is transmitted, the "own" bit is not cleared. As a result, most
of the implementations of is_tx_desc_closed() will return false. This causes
the TX interrupt service routine to quit early and leak an IOMMU mapping.
By testing for a beacon before calling the above routine, this leakage is
prevented.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>