rtlwifi_new: rtl8723be: Add new parameter "ant_sel"
A number of new laptops have only a single antenna. If the EFUSE setting
for hwinfo[EEPROM_RF_BT_SETTING_8723B] does not correspond to the connector
which has the connected antenna, the signal is very weak. This commit
adds a module parameter that allows the user to override the EFUSE setting.

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>