rtlwifi_new: Update to the latest rewrite now being submitted to mainline
Realtek has rewritten the BT coexistence routines. The new codes are being
submitted to mainline, and should be included in kernel 4.11. All of the
changes are included here. There is also new firmware for the RTL8723BE
and RTL8821AE.

Note that I have run the new code for about 5 days using the RTL8723BE
without any power-save options, and without any connection drops. The
new code is a definite improvement.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>