rtlwifi: Use dev_kfree_skb_irq instead of kfree_skb
It is not allowed to call kfree_skb() from hardware interrupt
context or with interrupts being disabled, spin_lock_irqsave()
make sure always in irq disable context. So the kfree_skb()
should be replaced with dev_kfree_skb_irq().

This is detected by Coccinelle semantic patch.

The original form failed to check for a NULL skb, which lead to
kernel panics. That problem has been fixed in mainline, and is
included in this commit.

Signed-off-by: Wei Yongjun <weiyongjun1@huawei.com>
Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>