rtlwifi_new: Fix test of RX descriptor
On four of the drivers, the wakeup test is performed twice on bit 31, rather
than once each on bits 29 and 31.

Thanks to Valerii Zapodovnikov for alerting me to this error.

Note that the only effect is to output the wrong value in an debug error
message. It has no affect on normal operation.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>